+++ src/01_Version.js
@@ -12,7 +12,7 @@
 
 const VERSION = {
  major: 0,
- minor: 29,
+ minor: 30,
  patch: 0,
 
  /**
@@ -24,7 +24,7 @@ const VERSION = {
  },
 
  releaseDate: '2026-08-20',
- releaseName: 'v0.29.0 - Vuelve el residuo: los tres destinos tienen que sumar 100% de los ingresos',
+ releaseName: 'v0.30.0 - Ningun mes se proyecta con desahorro',
 
  /**
  * Changelog embebido (solo refleja el release vigente).